+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Added
 - Tag panel for images in New images task [#2919]
+- Repository section for Collection object match task [#2918]
 
 ### Changed
 - Improvements for taxon determinations in comprehensive task
@@ -24,6 +25,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#2899]: https://github.com/SpeciesFileGroup/taxonworks/issues/2899
 [#2912]: https://github.com/SpeciesFileGroup/taxonworks/issues/2912
 [#2904]: https://github.com/SpeciesFileGroup/taxonworks/issues/2904
+[#2918]: https://github.com/SpeciesFileGroup/taxonworks/issues/2918
 [#2919]: https://github.com/SpeciesFileGroup/taxonworks/issues/2919
 
 ## [0.24.2] - 2022-04-15